refactor: rename files using common pattern
